requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# External Services
pinecone-client==5.0.1
huggingface-hub==0.20.3
langchain==0.1.14
langchain-community==0.0.24
pymupdf==1.23.8
pypdf==4.0.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0

# Build requirements
setuptools==68.2.2
wheel==0.41.3

